Refactor World to use ITerrain
