+++ src/main/java/com/thebois/models/world/TerrainType.java
@@ -1,0 +1,12 @@
+package com.thebois.models.world;
+
+/**
+ * All the different types of terrain in the game.
+ */
+public enum TerrainType {
+    /**
+     * The default ground terrain.
+     */
+    GRASS
+
+}
