+++ src/main/java/com/thebois/models/world/package-info.java
@@ -2,4 +2,4 @@
  * Containing all variants of tile.
  */
 
-package com.thebois.models.tiles;
+package com.thebois.models.world;
