+++ src/test/java/com/thebois/models/WorldTests.java
@@ -2,7 +2,8 @@ package com.thebois.models;
 
 import org.junit.jupiter.api.Test;
 
-import com.thebois.models.tiles.ITile;
+import com.thebois.models.world.ITerrain;
+import com.thebois.models.world.World;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -11,11 +12,11 @@ public class WorldTests {
     @Test
     public void worldInitiated() {
         // Arrange
-        final ITile[][] testWorld = new ITile[2][2];
+        final ITerrain[][] testWorld = new ITerrain[2][2];
         final World world = new World(2);
 
         // Act
-        final ITile[][] worldMatrix = world.getWorld();
+        final ITerrain[][] worldMatrix = world.getWorld();
 
         // Assert
         assertThat(matrixEquals(worldMatrix, testWorld)).isTrue();
@@ -33,9 +34,9 @@ public class WorldTests {
         assertThat(worldObject).isEqualTo(null);
     }
 
-    private boolean matrixEquals(ITile[][] worldMatrix1, ITile[][]worldMatrix2) {
+    private boolean matrixEquals(ITerrain[][] worldMatrix1, ITerrain[][]worldMatrix2) {
         for (int y = 0; y < worldMatrix2.length; y++) {
-            final ITile[] row = worldMatrix2[y];
+            final ITerrain[] row = worldMatrix2[y];
             for (int x = 0; x < row.length; x++) {
                 if (worldMatrix1[y][x] != null) {
                     if (!worldMatrix1[y][x].equals(worldMatrix2[y][x])) return false;
